+++ src/test/unit/plugin-specs.ts
@@ -1,14 +1,21 @@
 import { AppiumOcrPlugin } from '../..'
-import { BaseDriver } from 'appium/driver'
 import { expect } from 'earljs'
+
+const fakeDriver = {
+    async getSettings() {
+        return {}
+    }
+}
 
 describe('AppiumOcrPlugin', function() {
     it('should ready a worker', async function() {
         expect(AppiumOcrPlugin).toBeDefined()
         const p = new AppiumOcrPlugin('foo')
         expect(p.isWorkerReady).toEqual(false)
-        await p.readyWorker(new BaseDriver())
+        await p.readyWorker(fakeDriver)
         expect(p.isWorkerReady).toEqual(true)
-        await p.worker.terminate()
+        if (p.worker) {
+            await p.worker.terminate()
+        }
     })
 })
